+++ tests/test_commands/test_forecast.py
@@ -2,12 +2,17 @@ from pathlib import Path
 from subprocess import run
 from tempfile import NamedTemporaryFile
 
+import hydra_slayer
 import numpy as np
 import pandas as pd
 import pytest
+from omegaconf import OmegaConf
 
+from etna.commands.forecast_command import ADDITIONAL_PIPELINE_PARAMETERS
 from etna.commands.forecast_command import compute_horizon
 from etna.commands.forecast_command import filter_forecast
+from etna.commands.forecast_command import update_horizon
+from etna.commands.utils import remove_params
 from etna.datasets import TSDataset
 
 
@@ -59,14 +64,16 @@ def base_forecast_with_folds_estimation_omegaconf_path():
     tmp.close()
 
 
-def test_dummy_run_with_exog(base_pipeline_yaml_path, base_timeseries_path, base_timeseries_exog_path):
+@pytest.mark.parametrize("pipeline_path_name", ("base_pipeline_yaml_path", "base_ensemble_yaml_path"))
+def test_dummy_run_with_exog(pipeline_path_name, base_timeseries_path, base_timeseries_exog_path, request):
     tmp_output = NamedTemporaryFile("w")
     tmp_output_path = Path(tmp_output.name)
+    pipeline_path = request.getfixturevalue(pipeline_path_name)
     run(
         [
             "etna",
             "forecast",
-            str(base_pipeline_yaml_path),
+            str(pipeline_path),
             str(base_timeseries_path),
             "D",
             str(tmp_output_path),
@@ -103,16 +110,18 @@ def test_dummy_run(base_pipeline_yaml_path, base_timeseries_path):
     assert len(df_output) == 2 * 4
 
 
+@pytest.mark.parametrize("pipeline_path_name", ("base_pipeline_yaml_path", "base_ensemble_yaml_path"))
 def test_run_with_predictive_intervals(
-    base_pipeline_yaml_path, base_timeseries_path, base_timeseries_exog_path, base_forecast_omegaconf_path
+    pipeline_path_name, base_timeseries_path, base_timeseries_exog_path, base_forecast_omegaconf_path, request
 ):
     tmp_output = NamedTemporaryFile("w")
     tmp_output_path = Path(tmp_output.name)
+    pipeline_path = request.getfixturevalue(pipeline_path_name)
     run(
         [
             "etna",
             "forecast",
-            str(base_pipeline_yaml_path),
+            str(pipeline_path),
             str(base_timeseries_path),
             "D",
             str(tmp_output_path),
@@ -213,24 +222,45 @@ def test_filter_forecast(forecast_params, expected, example_tsds):
 
 
 @pytest.mark.parametrize(
-    "model_pipeline",
-    [
-        "elementary_linear_model_pipeline",
-        "elementary_boosting_model_pipeline",
-    ],
+    "forecast_params,pipeline_path_name,expected",
+    (
+        ({"start_timestamp": "2020-04-10"}, "base_pipeline_with_context_size_yaml_path", 4),
+        ({"start_timestamp": "2020-04-12"}, "base_pipeline_with_context_size_yaml_path", 6),
+        ({"start_timestamp": "2020-04-11"}, "base_ensemble_yaml_path", 5),
+    ),
+)
+def test_update_horizon(pipeline_path_name, forecast_params, example_tsds, expected, request):
+    pipeline_path = request.getfixturevalue(pipeline_path_name)
+    pipeline_conf = OmegaConf.to_object(OmegaConf.load(pipeline_path))
+
+    update_horizon(pipeline_configs=pipeline_conf, forecast_params=forecast_params, tsdataset=example_tsds)
+
+    pipeline_conf = remove_params(params=pipeline_conf, to_remove=ADDITIONAL_PIPELINE_PARAMETERS)
+    pipeline = hydra_slayer.get_from_params(**pipeline_conf)
+
+    assert pipeline.horizon == expected
+
+
+@pytest.mark.parametrize(
+    "pipeline_path_name",
+    ("base_pipeline_with_context_size_yaml_path", "base_ensemble_yaml_path"),
 )
 def test_forecast_start_timestamp(
-    model_pipeline, base_timeseries_path, base_timeseries_exog_path, start_timestamp_forecast_omegaconf_path, request
+    pipeline_path_name,
+    base_timeseries_path,
+    base_timeseries_exog_path,
+    start_timestamp_forecast_omegaconf_path,
+    request,
 ):
     tmp_output = NamedTemporaryFile("w")
     tmp_output_path = Path(tmp_output.name)
-    model_pipeline = request.getfixturevalue(model_pipeline)
+    pipeline_path = request.getfixturevalue(pipeline_path_name)
 
     run(
         [
             "etna",
             "forecast",
-            str(model_pipeline),
+            str(pipeline_path),
             str(base_timeseries_path),
             "D",
             str(tmp_output_path),
@@ -240,24 +270,28 @@ def test_forecast_start_timestamp(
     )
     df_output = pd.read_csv(tmp_output_path)
 
-    assert len(df_output) == 3 * 2  # 3 predictions for 2 segments
+    assert len(df_output) == 4 * 2  # 4 predictions for 2 segments
     assert df_output["timestamp"].min() == "2021-09-10"  # start_timestamp
     assert not np.any(df_output.isna().values)
 
 
+@pytest.mark.parametrize("pipeline_path_name", ("base_pipeline_with_context_size_yaml_path", "base_ensemble_yaml_path"))
 def test_forecast_estimate_n_folds(
-    base_pipeline_with_context_size_yaml_path,
+    pipeline_path_name,
     base_forecast_with_folds_estimation_omegaconf_path,
     base_timeseries_path,
     base_timeseries_exog_path,
+    request,
 ):
     tmp_output = NamedTemporaryFile("w")
     tmp_output_path = Path(tmp_output.name)
+    pipeline_path = request.getfixturevalue(pipeline_path_name)
+
     run(
         [
             "etna",
             "forecast",
-            str(base_pipeline_with_context_size_yaml_path),
+            str(pipeline_path),
             str(base_timeseries_path),
             "D",
             str(tmp_output_path),